Crate tls_codec

RustCrypto: TLS Codec

crates.io	docs.rs

tls_codec

tls_codec_derive

This crate implements the TLS codec as defined in RFC 8446 as well as some extensions required by MLS.

With the derive feature TlsSerialize and TlsDeserialize can be derived.

The crate also provides the following data structures that implement TLS serialization/deserialization

u8, u16, u32, u64

TlsVecU8, TlsVecU16, TlsVecU32

SecretTlsVecU8, SecretTlsVecU16, SecretTlsVecU32 The same as the TlsVec* versions but it implements zeroize, requiring the elements to implement zeroize as well.

TlsSliceU8, TlsSliceU16, TlsSliceU32 are lightweight wrapper for slices that allow to serialize them without having to create a TlsVec*.

TlsByteSliceU8, TlsByteSliceU16, TlsByteSliceU32, and TlsByteVecU8, TlsByteVecU16, TlsByteVecU32 are provided with optimized implementations for byte vectors.

[u8; l], for l ∈ [1..128]

Serialize for Option<T> where T: Serialize

Deserialize for Option<T> where T: Deserialize

Serialize for (T, U) and (T, U, V) where T, U, V implement Serialize`

Deserialize for (T, U) and (T, U, V) where T, U, V implement Deserialize`

Minimum Supported Rust Version

This crate requires Rust 1.74.0 at a minimum.

We may change the MSRV in the future, but it will be accompanied by a minor version bump.

License

Licensed under either of:

Apache License, Version 2.0

MIT license

at your option.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Usage

use tls_codec::{TlsVecU8, Serialize, Deserialize};
let mut b = &[1u8, 4, 77, 88, 1, 99] as &[u8];

let a = u8::tls_deserialize(&mut b).expect("Unable to tls_deserialize");
assert_eq!(1, a);
println!("b: {:?}", b);
let v = TlsVecU8::<u8>::tls_deserialize(&mut b).expect("Unable to tls_deserialize");
assert_eq!(&[77, 88, 1, 99], v.as_slice());

Modules

vlen

Structs

SecretTlsVecU8

SecretTlsVecU16

SecretTlsVecU24

SecretTlsVecU32

SecretVLBytes

A wrapper struct around VLBytes that implements ZeroizeOnDrop. It behaves just like VLBytes, except that it doesn’t allow conversion into a Vec<u8>.

TlsByteSliceU8

TlsByteSliceU16

TlsByteSliceU24

TlsByteSliceU32

TlsByteVecU8

TlsByteVecU16

TlsByteVecU24

TlsByteVecU32

TlsSliceU8

TlsSliceU16

TlsSliceU24

TlsSliceU32

TlsVecU8

TlsVecU16

TlsVecU24

TlsVecU32

U24

A 3 byte wide unsigned integer type as defined in RFC 5246.

VLByteSlice

VLBytes

Variable-length encoded byte vectors. Use this struct if bytes are encoded. This is faster than the generic version.

Enums

Error

Errors that are thrown by this crate.

Traits

Deserialize

The Deserialize trait defines functions to deserialize a byte slice to a struct or enum.

DeserializeBytes

The DeserializeBytes trait defines functions to deserialize a byte slice to a struct or enum. In contrast to Deserialize, this trait operates directly on byte slices and can return any remaining bytes.

Serialize

The Serialize trait provides functions to serialize a struct or enum.

SerializeBytes

The SerializeBytes trait provides a function to serialize a struct or enum.

Size

The Size trait needs to be implemented by any struct that should be efficiently serialized. This allows to collect the length of a serialized structure before allocating memory.

Derive Macros

TlsDeserialize

TlsDeserializeBytes

TlsSerialize

TlsSerializeBytes

TlsSize