Struct tls_codec::TlsVecU32

pub struct TlsVecU32<T: Serialize + Deserialize + Clone + PartialEq + TlsSize> { /* fields omitted */ }

Implementations

impl<T: Serialize + Deserialize + Clone + PartialEq + TlsSize> TlsVecU32<T>

pub fn new(vec: Vec<T>) -> Self

Create a new TlsVec from a Rust Vec.

pub fn from_slice(slice: &[T]) -> Self

Create a new TlsVec from a slice.

pub fn len(&self) -> usize

Get the length of the vector.

pub fn as_slice(&self) -> &[T]

Get a slice to the raw vector.

pub fn to_vec(&self) -> Vec<T>

Get a copy of the underlying vector.

pub fn is_empty(&self) -> bool

Check if the vector is empty.

pub fn into_vec(self) -> Vec<T>

Get the underlying vector and consume this.

pub fn push(&mut self, value: T)

Add an element to this.

pub fn pop(&mut self) -> Option<T>

Remove the last element.

pub fn get(&self, index: usize) -> Option<&T>

Returns a reference to an element or subslice depending on the type of index. XXX: implement SliceIndex instead

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator over the slice.

pub fn retain<F>(&mut self, f: F) where
F: FnMut(&T) -> bool,

Retains only the elements specified by the predicate.

pub fn len_len() -> usize

Get the number of bytes used for the length encoding.