tls_codec

Trait Serialize

pub trait Serialize: Size {
    // Required method
    fn tls_serialize<W: Write>(&self, writer: &mut W) -> Result<usize, Error>;

    // Provided method
    fn tls_serialize_detached(&self) -> Result<Vec<u8>, Error> { ... }
}

The Serialize trait provides functions to serialize a struct or enum.

The trait provides two functions:

tls_serialize that takes a buffer to write the serialization to

tls_serialize_detached that returns a byte vector

Required Methods

fn tls_serialize<W: Write>(&self, writer: &mut W) -> Result<usize, Error>

Serialize self and write it to the writer. The function returns the number of bytes written to writer.

Provided Methods

fn tls_serialize_detached(&self) -> Result<Vec<u8>, Error>

Serialize self and return it as a byte vector.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.