Crate tls_codec

TLS Codec

This crate implements the TLS codec as defined in RFC 8446 as well as some extensions required by MLS.

With the feature derive TlsSerialize and TlsDeserialize can be derived.

This crate provides the following data structures that implement TLS serialization

u8, u16, u32, u64

TlsVecU8, TlsVecU16, TlsVecU32

[u8; 2], [u8; 4], [u8; 8], [u8; 16], [u8; 32], [u8; 64]

If the hpke feature is enabled, the TLS codec is implemented for HPKEPublicKeys from hpke-rs.

Modules

arrays
Implement the TLS codec for some byte arrays.

hpke
TLS Codec for HPKE primitives

primitives
Codec implementations for unsigned integer primitives.

tls_vec
A vector with a length field for TLS serialisation. Use this for any vector that is serialised.

Structs

Cursor
The Cursor is a helper used to read byte slices when deserializing. Note that this shouldn’t be needed in most cases and tls_deserialize_detached can be used instead.

TlsVecU8

TlsVecU16

TlsVecU32

Enums

Error
Errors that are thrown by this crate.

Traits

Deserialize
The Deserialize trait provides functions to deserialize a byte slice to a struct or enum.

Serialize
The Serialize trait provides functions to serialize a struct or enum.

TlsSize
The TlsSize trait needs to be implemented by any struct that should be efficiently serialized. This allows to collect the length of a serialized structure before allocating memory.

Derive Macros

TlsDeserialize

TlsSerialize