Trait tls_codec::Serialize

pub trait Serialize: Size {
    fn tls_serialize<W: Write>(&self, writer: &mut W) -> Result<usize, Error>;

    fn tls_serialize_detached(&self) -> Result<Vec<u8>, Error> { ... }
}

The Serialize trait provides functions to serialize a struct or enum.

The trait provides two functions:

tls_serialize that takes a buffer to write the serialization to

tls_serialize_detached that returns a byte vector

Required methods

fn tls_serialize<W: Write>(&self, writer: &mut W) -> Result<usize, Error>

Serialize self and write it to the writer. The function returns the number of bytes written to writer.

Loading content...
Provided methods

fn tls_serialize_detached(&self) -> Result<Vec<u8>, Error>

Serialize self and return it as a byte vector.

Loading content...