Trait tls_codec::TlsSize

pub trait TlsSize {
    fn serialized_len(&self) -> usize;
}

The TlsSize trait needs to be implemented by any struct that should be efficiently serialized. This allows to collect the length of a serialized structure before allocating memory.

Required methods

fn serialized_len(&self) -> usize

Loading content...